+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - **Breaking**: Changed `HolodexClient::searchComments` return type to `PaginatedResult<VideoWithComments>`
 - **Breaking**: Changed `HolodexClient::getVideos` parameters to an optional `VideoFilter`
 - **Breaking**: Changed `HolodexClient::getLiveVideos` parameters to an optional `VideoFilter`
+- **Breaking**: Changed `HolodexClient::getChannels` parameters to an optional `ChannelFilter`
 - **Breaking**: Removed `VideoWithCommentsList`
 - **Breaking**: Renamed `client.dart` to `holodex_client.dart`. Technically a breaking change, but only if
 implementation was imported.